Sticky Notes
We can paste our sticky notes on the screen with this app!, created functions to pick random value for margins, note colors and rotation of notes to give a realistic look. Positioned the textarea such that it will always appear on top of the notes in the center of the screen.